extract run tab component
